feat: add react-hot-toast
* refactor: AppFormConfigProvider out of _app.tsx
