+++ package.json
@@ -20,7 +20,8 @@
     "next-auth": "^4.20.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.43.7"
+    "react-hook-form": "^7.43.7",
+    "react-hot-toast": "2.4.1"
   },
   "devDependencies": {
     "@swc/core": "^1.3.52",
